Add `Perform::terminated` function
This patch is a rework of the partial processing patch in an attempt to
provide an identical clean API while still allowing for arbitrary
terminated of the parser for partial synchronized update processing.

Instead of returning values using the dispatch functions, a separate
`Perform::terminated` function is added which is queried whenever the
new `advance_until_terminated` function is called. The normal `advance`
function stays unchanged.

While the `advance` function could be implemented using the
`advance_until_terminated` function, this seems like it would just add
an unnecessary performance overhead. So since the function is pretty
small its contents are just duplicated instead.